move plugin.Execute in internal package
